pyproject.toml
[tool.poetry]
name = "translatorbot"
version = "0.1.0"
description = ""
authors = ["beerpiss <lacvtg.a1.2023@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
discord-py = "^2.2.2"
googletrans = "3.1.0a0"
python-dotenv = "^1.0.0"
langcodes = "^3.3.0"
language-data = "^1.1"
lingua-language-detector = "^1.3.2"
spacy = "^3.5.1"
spacy-language-detection = "^0.2.1"
pydrive2 = "^1.15.2"


[tool.poetry.group.dev.dependencies]
mypy = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
